% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openApp.R
\name{openApp}
\alias{openApp}
\title{openApp}
\usage{
openApp(dir = "mini", ...)
}
\arguments{
\item{dir}{character: name of the app}

\item{...}{further parameter given to \code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}}}
}
\value{
invisibly the result of \code{navigateToFile}
}
\description{
Opens an example Shiny app into the RStudio editor.
}
\examples{
if (interactive()) run() # runs the minimal example
}
