% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetTimer.R
\name{resetTimer}
\alias{resetTimer}
\title{resetTimer}
\usage{
resetTimer(inputId)
}
\arguments{
\item{inputId}{character: input slot}
}
\value{
nothing
}
\description{
Simulates a button press for \code{Reset}.
}
\examples{
if (interactive()) vignette("shinyDTC")
}
