% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAppx.R
\name{runAppx}
\alias{runAppx}
\title{runAppx}
\usage{
runAppx(example = "mini", x = NULL, ...)
}
\arguments{
\item{example}{character: name of the example to run, or NA to list the available examples}

\item{x}{data frame: data to pass to the app}

\item{...}{further parameter given to \code{\link[shiny:runExample]{shiny::runExample()}}}
}
\value{
invisibly the (modified) data set \code{x}
}
\description{
Runs an example Shiny app via \code{\link[shiny:runExample]{shiny::runExample()}}.
}
\details{
Note that for running a Shiny app more libraries might be necessary.
}
\examples{
if (interactive()) runAppx() # runs the minimal example
}
