% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepTimer.R
\name{stepTimer}
\alias{stepTimer}
\title{stepTimer}
\usage{
stepTimer(inputId)
}
\arguments{
\item{inputId}{character: input slot}
}
\value{
nothing
}
\description{
Simulates a button press on \code{Step}.
}
\examples{
if (interactive()) vignette("shinyDTC")
}
