% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{timer}
\usage{
timer(inputId, input, session, ...)
}
\arguments{
\item{inputId}{character: input slot for a timer}

\item{input}{an inpur object}

\item{session}{a session object}

\item{...}{further parameters for \code{\link[shiny:reactive]{shiny::reactive()}}}
}
\value{
a reactive function which returns a counter
}
\description{
A reactive function.
}
\examples{
if (interactive()) vignette("shinyDTC")
}
