% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timerInput.R
\name{timerInput}
\alias{timerInput}
\title{timerInput}
\usage{
timerInput(inputId, step1 = list(), reset = list(), ...)
}
\arguments{
\item{inputId}{character: input slot}

\item{step1}{list: list of parameters for the left (step) button}

\item{reset}{list: list of parameters for the right (reset) button}

\item{...}{parameters to the slider}
}
\value{
a widget to timer
}
\description{
Constructs a timer widget with a slider and two buttons.
}
\examples{
if (interactive()) vignette("shinyDTC")
}
