% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTimerInput.R
\name{updateTimerInput}
\alias{updateTimerInput}
\title{updateTimerInput}
\usage{
updateTimerInput(
  session = getDefaultReactiveDomain(),
  inputId,
  step1 = list(),
  reset = list(),
  ...
)
}
\arguments{
\item{session}{session object}

\item{inputId}{id of the input object}

\item{step1}{modified values for \code{Step} button, see \code{\link[shiny:updateActionButton]{shiny::updateActionButton()}}}

\item{reset}{modified values for \code{Reset} button, see \code{\link[shiny:updateActionButton]{shiny::updateActionButton()}}}

\item{...}{modified values for \code{Speed} slider, see \code{\link[shiny:updateSliderInput]{shiny::updateSliderInput()}}}
}
\value{
nothing
}
\description{
updateTimerInput
}
\examples{
if (interactive()) vignette("shinyDTC")
}
